move some utils to Geo module
